+++ service/serving/model.go
@@ -33,6 +33,202 @@ func (s Ai21LabsConfig) MarshalJSON() ([]byte, error) {
 	return marshal.Marshal(s)
 }
 
+type AiGatewayConfig struct {
+	// Configuration for AI Guardrails to prevent unwanted data and unsafe data
+	// in requests and responses.
+	Guardrails *AiGatewayGuardrails `json:"guardrails,omitempty"`
+	// Configuration for payload logging using inference tables. Use these
+	// tables to monitor and audit data being sent to and received from model
+	// APIs and to improve model quality.
+	InferenceTableConfig *AiGatewayInferenceTableConfig `json:"inference_table_config,omitempty"`
+	// Configuration for rate limits which can be set to limit endpoint traffic.
+	RateLimits []AiGatewayRateLimit `json:"rate_limits,omitempty"`
+	// Configuration to enable usage tracking using system tables. These tables
+	// allow you to monitor operational usage on endpoints and their associated
+	// costs.
+	UsageTrackingConfig *AiGatewayUsageTrackingConfig `json:"usage_tracking_config,omitempty"`
+}
+
+type AiGatewayGuardrailParameters struct {
+	// List of invalid keywords. AI guardrail uses keyword or string matching to
+	// decide if the keyword exists in the request or response content.
+	InvalidKeywords []string `json:"invalid_keywords,omitempty"`
+	// Configuration for guardrail PII filter.
+	Pii *AiGatewayGuardrailPiiBehavior `json:"pii,omitempty"`
+	// Indicates whether the safety filter is enabled.
+	Safety bool `json:"safety,omitempty"`
+	// The list of allowed topics. Given a chat request, this guardrail flags
+	// the request if its topic is not in the allowed topics.
+	ValidTopics []string `json:"valid_topics,omitempty"`
+
+	ForceSendFields []string `json:"-"`
+}
+
+func (s *AiGatewayGuardrailParameters) UnmarshalJSON(b []byte) error {
+	return marshal.Unmarshal(b, s)
+}
+
+func (s AiGatewayGuardrailParameters) MarshalJSON() ([]byte, error) {
+	return marshal.Marshal(s)
+}
+
+type AiGatewayGuardrailPiiBehavior struct {
+	// Behavior for PII filter. Currently only 'BLOCK' is supported. If 'BLOCK'
+	// is set for the input guardrail and the request contains PII, the request
+	// is not sent to the model server and 400 status code is returned; if
+	// 'BLOCK' is set for the output guardrail and the model response contains
+	// PII, the PII info in the response is redacted and 400 status code is
+	// returned.
+	Behavior AiGatewayGuardrailPiiBehaviorBehavior `json:"behavior"`
+}
+
+// Behavior for PII filter. Currently only 'BLOCK' is supported. If 'BLOCK' is
+// set for the input guardrail and the request contains PII, the request is not
+// sent to the model server and 400 status code is returned; if 'BLOCK' is set
+// for the output guardrail and the model response contains PII, the PII info in
+// the response is redacted and 400 status code is returned.
+type AiGatewayGuardrailPiiBehaviorBehavior string
+
+const AiGatewayGuardrailPiiBehaviorBehaviorBlock AiGatewayGuardrailPiiBehaviorBehavior = `BLOCK`
+
+const AiGatewayGuardrailPiiBehaviorBehaviorNone AiGatewayGuardrailPiiBehaviorBehavior = `NONE`
+
+// String representation for [fmt.Print]
+func (f *AiGatewayGuardrailPiiBehaviorBehavior) String() string {
+	return string(*f)
+}
+
+// Set raw string value and validate it against allowed values
+func (f *AiGatewayGuardrailPiiBehaviorBehavior) Set(v string) error {
+	switch v {
+	case `BLOCK`, `NONE`:
+		*f = AiGatewayGuardrailPiiBehaviorBehavior(v)
+		return nil
+	default:
+		return fmt.Errorf(`value "%s" is not one of "BLOCK", "NONE"`, v)
+	}
+}
+
+// Type always returns AiGatewayGuardrailPiiBehaviorBehavior to satisfy [pflag.Value] interface
+func (f *AiGatewayGuardrailPiiBehaviorBehavior) Type() string {
+	return "AiGatewayGuardrailPiiBehaviorBehavior"
+}
+
+type AiGatewayGuardrails struct {
+	// Configuration for input guardrail filters.
+	Input *AiGatewayGuardrailParameters `json:"input,omitempty"`
+	// Configuration for output guardrail filters.
+	Output *AiGatewayGuardrailParameters `json:"output,omitempty"`
+}
+
+type AiGatewayInferenceTableConfig struct {
+	// The name of the catalog in Unity Catalog. Required when enabling
+	// inference tables. NOTE: On update, you have to disable inference table
+	// first in order to change the catalog name.
+	CatalogName string `json:"catalog_name,omitempty"`
+	// Indicates whether the inference table is enabled.
+	Enabled bool `json:"enabled,omitempty"`
+	// The name of the schema in Unity Catalog. Required when enabling inference
+	// tables. NOTE: On update, you have to disable inference table first in
+	// order to change the schema name.
+	SchemaName string `json:"schema_name,omitempty"`
+	// The prefix of the table in Unity Catalog. NOTE: On update, you have to
+	// disable inference table first in order to change the prefix name.
+	TableNamePrefix string `json:"table_name_prefix,omitempty"`
+
+	ForceSendFields []string `json:"-"`
+}
+
+func (s *AiGatewayInferenceTableConfig) UnmarshalJSON(b []byte) error {
+	return marshal.Unmarshal(b, s)
+}
+
+func (s AiGatewayInferenceTableConfig) MarshalJSON() ([]byte, error) {
+	return marshal.Marshal(s)
+}
+
+type AiGatewayRateLimit struct {
+	// Used to specify how many calls are allowed for a key within the
+	// renewal_period.
+	Calls int `json:"calls"`
+	// Key field for a rate limit. Currently, only 'user' and 'endpoint' are
+	// supported, with 'endpoint' being the default if not specified.
+	Key AiGatewayRateLimitKey `json:"key,omitempty"`
+	// Renewal period field for a rate limit. Currently, only 'minute' is
+	// supported.
+	RenewalPeriod AiGatewayRateLimitRenewalPeriod `json:"renewal_period"`
+}
+
+// Key field for a rate limit. Currently, only 'user' and 'endpoint' are
+// supported, with 'endpoint' being the default if not specified.
+type AiGatewayRateLimitKey string
+
+const AiGatewayRateLimitKeyEndpoint AiGatewayRateLimitKey = `endpoint`
+
+const AiGatewayRateLimitKeyUser AiGatewayRateLimitKey = `user`
+
+// String representation for [fmt.Print]
+func (f *AiGatewayRateLimitKey) String() string {
+	return string(*f)
+}
+
+// Set raw string value and validate it against allowed values
+func (f *AiGatewayRateLimitKey) Set(v string) error {
+	switch v {
+	case `endpoint`, `user`:
+		*f = AiGatewayRateLimitKey(v)
+		return nil
+	default:
+		return fmt.Errorf(`value "%s" is not one of "endpoint", "user"`, v)
+	}
+}
+
+// Type always returns AiGatewayRateLimitKey to satisfy [pflag.Value] interface
+func (f *AiGatewayRateLimitKey) Type() string {
+	return "AiGatewayRateLimitKey"
+}
+
+// Renewal period field for a rate limit. Currently, only 'minute' is supported.
+type AiGatewayRateLimitRenewalPeriod string
+
+const AiGatewayRateLimitRenewalPeriodMinute AiGatewayRateLimitRenewalPeriod = `minute`
+
+// String representation for [fmt.Print]
+func (f *AiGatewayRateLimitRenewalPeriod) String() string {
+	return string(*f)
+}
+
+// Set raw string value and validate it against allowed values
+func (f *AiGatewayRateLimitRenewalPeriod) Set(v string) error {
+	switch v {
+	case `minute`:
+		*f = AiGatewayRateLimitRenewalPeriod(v)
+		return nil
+	default:
+		return fmt.Errorf(`value "%s" is not one of "minute"`, v)
+	}
+}
+
+// Type always returns AiGatewayRateLimitRenewalPeriod to satisfy [pflag.Value] interface
+func (f *AiGatewayRateLimitRenewalPeriod) Type() string {
+	return "AiGatewayRateLimitRenewalPeriod"
+}
+
+type AiGatewayUsageTrackingConfig struct {
+	// Whether to enable usage tracking.
+	Enabled bool `json:"enabled,omitempty"`
+
+	ForceSendFields []string `json:"-"`
+}
+
+func (s *AiGatewayUsageTrackingConfig) UnmarshalJSON(b []byte) error {
+	return marshal.Unmarshal(b, s)
+}
+
+func (s AiGatewayUsageTrackingConfig) MarshalJSON() ([]byte, error) {
+	return marshal.Marshal(s)
+}
+
 type AmazonBedrockConfig struct {
 	// The Databricks secret key reference for an AWS access key ID with
 	// permissions to interact with Bedrock services. If you prefer to paste
@@ -273,14 +469,17 @@ func (s CohereConfig) MarshalJSON() ([]byte, error) {
 }
 
 type CreateServingEndpoint struct {
+	// The AI Gateway configuration for the serving endpoint. NOTE: only
+	// external model endpoints are supported as of now.
+	AiGateway *AiGatewayConfig `json:"ai_gateway,omitempty"`
 	// The core config of the serving endpoint.
 	Config EndpointCoreConfigInput `json:"config"`
 	// The name of the serving endpoint. This field is required and must be
 	// unique across a Databricks workspace. An endpoint name can consist of
 	// alphanumeric characters, dashes, and underscores.
 	Name string `json:"name"`
-	// Rate limits to be applied to the serving endpoint. NOTE: only external
-	// and foundation model endpoints are supported as of now.
+	// Rate limits to be applied to the serving endpoint. NOTE: this field is
+	// deprecated, please use AI Gateway to manage rate limits.
 	RateLimits []RateLimit `json:"rate_limits,omitempty"`
 	// Enable route optimization for the serving endpoint.
 	RouteOptimized bool `json:"route_optimized,omitempty"`
@@ -901,6 +1100,42 @@ func (s PayloadTable) MarshalJSON() ([]byte, error) {
 	return marshal.Marshal(s)
 }
 
+// Update AI Gateway of a serving endpoint
+type PutAiGatewayRequest struct {
+	// Configuration for AI Guardrails to prevent unwanted data and unsafe data
+	// in requests and responses.
+	Guardrails *AiGatewayGuardrails `json:"guardrails,omitempty"`
+	// Configuration for payload logging using inference tables. Use these
+	// tables to monitor and audit data being sent to and received from model
+	// APIs and to improve model quality.
+	InferenceTableConfig *AiGatewayInferenceTableConfig `json:"inference_table_config,omitempty"`
+	// The name of the serving endpoint whose AI Gateway is being updated. This
+	// field is required.
+	Name string `json:"-" url:"-"`
+	// Configuration for rate limits which can be set to limit endpoint traffic.
+	RateLimits []AiGatewayRateLimit `json:"rate_limits,omitempty"`
+	// Configuration to enable usage tracking using system tables. These tables
+	// allow you to monitor operational usage on endpoints and their associated
+	// costs.
+	UsageTrackingConfig *AiGatewayUsageTrackingConfig `json:"usage_tracking_config,omitempty"`
+}
+
+type PutAiGatewayResponse struct {
+	// Configuration for AI Guardrails to prevent unwanted data and unsafe data
+	// in requests and responses.
+	Guardrails *AiGatewayGuardrails `json:"guardrails,omitempty"`
+	// Configuration for payload logging using inference tables. Use these
+	// tables to monitor and audit data being sent to and received from model
+	// APIs and to improve model quality .
+	InferenceTableConfig *AiGatewayInferenceTableConfig `json:"inference_table_config,omitempty"`
+	// Configuration for rate limits which can be set to limit endpoint traffic.
+	RateLimits []AiGatewayRateLimit `json:"rate_limits,omitempty"`
+	// Configuration to enable usage tracking using system tables. These tables
+	// allow you to monitor operational usage on endpoints and their associated
+	// costs.
+	UsageTrackingConfig *AiGatewayUsageTrackingConfig `json:"usage_tracking_config,omitempty"`
+}
+
 // Update rate limits of a serving endpoint
 type PutRequest struct {
 	// The name of the serving endpoint whose rate limits are being updated.
@@ -1313,6 +1548,10 @@ type ServedModelInput struct {
 	// ARN of the instance profile that the served model will use to access AWS
 	// resources.
 	InstanceProfileArn string `json:"instance_profile_arn,omitempty"`
+	// The maximum tokens per second that the endpoint can scale up to.
+	MaxProvisionedThroughput int `json:"max_provisioned_throughput,omitempty"`
+	// The minimum tokens per second that the endpoint can scale down to.
+	MinProvisionedThroughput int `json:"min_provisioned_throughput,omitempty"`
 	// The name of the model in Databricks Model Registry to be served or if the
 	// model resides in Unity Catalog, the full name of model, in the form of
 	// __catalog_name__.__schema_name__.__model_name__.
@@ -1335,7 +1574,7 @@ type ServedModelInput struct {
 	// "Medium" (8 - 16 provisioned concurrency), and "Large" (16 - 64
 	// provisioned concurrency). If scale-to-zero is enabled, the lower bound of
 	// the provisioned concurrency for each workload size will be 0.
-	WorkloadSize ServedModelInputWorkloadSize `json:"workload_size"`
+	WorkloadSize ServedModelInputWorkloadSize `json:"workload_size,omitempty"`
 	// The workload type of the served model. The workload type selects which
 	// type of compute to use in the endpoint. The default value for this
 	// parameter is "CPU". For deep learning workloads, GPU acceleration is
@@ -1588,6 +1827,9 @@ type ServerLogsResponse struct {
 }
 
 type ServingEndpoint struct {
+	// The AI Gateway configuration for the serving endpoint. NOTE: Only
+	// external model endpoints are currently supported.
+	AiGateway *AiGatewayConfig `json:"ai_gateway,omitempty"`
 	// The config that is currently being served by the endpoint.
 	Config *EndpointCoreConfigSummary `json:"config,omitempty"`
 	// The timestamp when the endpoint was created in Unix time.
@@ -1664,6 +1906,9 @@ func (s ServingEndpointAccessControlResponse) MarshalJSON() ([]byte, error) {
 }
 
 type ServingEndpointDetailed struct {
+	// The AI Gateway configuration for the serving endpoint. NOTE: Only
+	// external model endpoints are currently supported.
+	AiGateway *AiGatewayConfig `json:"ai_gateway,omitempty"`
 	// The config that is currently being served by the endpoint.
 	Config *EndpointCoreConfigOutput `json:"config,omitempty"`
 	// The timestamp when the endpoint was created in Unix time.